Fix integer signedness in Vulkan ConvertToMeshOutputCompute
Mesa's SPIR-V parser as well as spirv-val reject shaders generated by this
function because a sint32 {vertex,instance}{Index,Base} is stored in a
uint32 variable.

I'm not familiar enough with the code to be sure if this change is complete
or even the right way to do it -- I was a bit surprised to see that the
intermediate calculations are with signed integers in the first place.
Presumably it makes a difference for div/mod, although I'd have thought it's
more natural to just do all calculations as uint32.

This patch is simply the most conservative change that allowed me to debug
the trace that I needed to debug.

Signed-off-by: Nicolai Hähnle <nicolai.haehnle@amd.com>